+++ runtime/moonbeam/src/runtime_params.rs
@@ -16,11 +16,47 @@
 
 //! Dynamic runtime parameters for moonbeam.
 
-use moonbeam_runtime_common::gen_runtime_params;
+use crate::{currency, Runtime};
+use frame_support::dynamic_params::{dynamic_pallet_params, dynamic_params};
+use moonbeam_runtime_common::expose_u128_get;
+use moonbeam_runtime_common::types::BoundedU128;
+use sp_runtime::Perbill;
 
-gen_runtime_params!(
-	UNIT: GLMR,
-	RuntimeConfig_TreasuryProportion: Perbill::from_percent(20),
-	PalletRandomness_Deposit:
-		BoundedU128::const_new::<{ 1 * currency::GLMR * currency::SUPPLY_FACTOR }>(),
+#[dynamic_params(RuntimeParameters, pallet_parameters::Parameters::<Runtime>)]
+pub mod dynamic_params {
+	use super::*;
+	#[dynamic_pallet_params]
+	#[codec(index = 0)]
+	pub mod runtime_config {
+		// for fees, 80% are burned, 20% to the treasury
+		#[codec(index = 0)]
+		pub static FeesTreasuryProportion: Perbill = Perbill::from_percent(20);
+	}
+
+	#[dynamic_pallet_params]
+	#[codec(index = 1)]
+	pub mod pallet_randomness {
+		#[codec(index = 0)]
+		pub static Deposit: BoundedU128<
+			{ 1 * currency::GLMR * currency::SUPPLY_FACTOR },
+			{ 1_000 * currency::GLMR * currency::SUPPLY_FACTOR },
+		> = BoundedU128::const_new::<{ 1 * currency::GLMR * currency::SUPPLY_FACTOR }>();
+	}
+}
+
+expose_u128_get!(
+	PalletRandomnessDepositU128,
+	dynamic_params::pallet_randomness::Deposit
 );
+
+#[cfg(feature = "runtime-benchmarks")]
+impl Default for RuntimeParameters {
+	fn default() -> Self {
+		RuntimeParameters::RuntimeConfig(
+			dynamic_params::runtime_config::Parameters::FeesTreasuryProportion(
+				dynamic_params::runtime_config::FeesTreasuryProportion,
+				Some(Perbill::from_percent(20)),
+			),
+		)
+	}
+}
