+++ runtime/moonriver/src/lib.rs
@@ -1435,6 +1435,7 @@ construct_runtime! {
 		ProxyGenesisCompanion: pallet_proxy_genesis_companion::{Pallet, Config<T>} = 35,
 		Multisig: pallet_multisig::{Pallet, Call, Storage, Event<T>} = 36,
 		MoonbeamLazyMigrations: pallet_moonbeam_lazy_migrations::{Pallet, Call, Storage} = 37,
+		Parameters: pallet_parameters = 38,
 
 		// Sudo was previously index 40
 
@@ -1489,8 +1490,6 @@ construct_runtime! {
 
 		// Randomness
 		Randomness: pallet_randomness::{Pallet, Call, Storage, Event<T>, Inherent} = 120,
-
-		Parameters: pallet_parameters = 121,
 	}
 }
